refactor context menu creation
